pyproject.toml
[project]
name = "stitch-worker"
version = "0.1.0"
description = "A CDK app with 5 Lambda functions and 5 SQS queues"
authors = [
    {name = "Jason DeCorte", email = "jason@stitchstudio.ai"}
]
dependencies = [
    "aws-cdk-lib>=2.189.1",
    "constructs>=10.4.2"
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stitch_worker"]
